decouple job prototype
