+++ lib/http/routes/json.js
@@ -158,7 +158,7 @@ exports.createJob = function( req, res ) {
   function _create( args, next ) {
     if( !args.type ) return next({ error: 'Must provide job type' }, null, 400);
 
-    var job     = new Job.Job(args.type, args.data || {});
+    var job     = queue.createJob(args.type, args.data || {});
     var options = args.options || {};
     if( options.attempts ) job.attempts(parseInt(options.attempts));
     if( options.priority ) job.priority(options.priority);
