+++ lib/queueEvents.js
