+++ test/jsonapi.js
@@ -32,7 +32,7 @@ function jobsPopulate( count ) {
 }
 
 
-describe( 'JSON API', function () {
+describe.only( 'JSON API', function () {
   var scope = {};
